initial working version

+++ edit_tds_datasets.py
@@ -5,8 +5,14 @@ import sys
 import argparse
 import csv
 import glob
+import shutil
 
 def main(args):
+    '''Parses CSV_FILE for records containing a reference designator, telemetry type
+    and stream name.  Each record is used to create and validate a THREDDS stream location
+    (.ncml and .nc files) and the list of files is printed to STDOUT.  Additional
+    options are available to copy, move or delete the found files and resulting
+    degenerate directory tree.'''
     
     _OOI_ARRAYS = {'CP' : 'Coastal_Pioneer',
         'CE' : 'Coastal_Endurance',
@@ -22,7 +28,7 @@ def main(args):
         'stream',
         'telemetry']
         
-    status = 0
+    status = 1
     
     TDS_NC_ROOT = args.tdsroot
     if not TDS_NC_ROOT:
@@ -39,7 +45,8 @@ def main(args):
         sys.stderr.flush()
         return 1
     
-    # If delete or copy was specified, check args.location to ensure it is valid
+    # If move or copy was specified, check args.location to ensure it is valid
+    location = args.location
     if args.move or args.copy:
         if not args.location:
             sys.stderr.write('No THREDDS destination specified\n')
@@ -47,15 +54,8 @@ def main(args):
         if not os.path.exists(args.location):
             sys.stderr.write('Invalid destination specified: {:s}\n'.format(args.location))
             return status
-            
-    # Open args.csv_file for reading    
+                
     csv_file = args.csv_file
-    #fid = open(csv_file, 'r')
-    #csv_reader = csv.reader(fid)
-    ## First row is the header
-    #cols = csv_reader.next()
-    #fid.close()
-    
     # convert the csv file rows to a dictionary
     streams = csv2json(csv_file)
     if not streams:
@@ -101,6 +101,7 @@ def main(args):
             continue
             
         if not args.delete and not args.copy and not args.move:
+            sys.stdout.write('No file operations will be performed\n')
             # Print the list of files found, then skip since we're not operating on
             # these files
             for f in f_contents:
@@ -108,14 +109,78 @@ def main(args):
                 
             continue
             
-        if args.delete:
+        elif location and not os.path.exists(location):
+            # Validate the specified destination root directory
+            sys.stderr.write('Invalid destination root: {:s}\n'.format(location))
+            return status
+           
+        # Copy, move or delete the stream tree if specified on the command line 
+        if args.copy:
+            sys.stdout.write('Copying THREDDS stream to {:s}\n'.format(location))
+            
+            # Create the rel_path under location
+            new_location = os.path.join(location, rel_path)
+            if not os.path.exists(new_location):
+                try:
+                    os.makedirs(new_location)
+                except OSError as e:
+                    sys.stderr.write('{:s}: {:s}\n'.format(e.strerror, new_location))
+                    continue
+                    
+            for f in f_contents:
+                # Copy the file only if it doesn't already exist
+                dest_nc = os.path.join(new_location, os.path.basename(f))
+                if os.path.exists(dest_nc):
+                    sys.stderr.write('Destination file already exists: {:s}\n'.format(dest_nc))
+                    continue
+                try:
+                    sys.stdout.write('Copying {:s}: {:s}\n'.format(os.path.basename(f), new_location))
+                    shutil.copy(f, new_location)
+                except IOError as e:
+                    sys.stderr.write('{:s}: {:s}\n'.format(e.strerror, f))
+                    continue
+                
+        elif args.move:
+            sys.stdout.write('Moving THREDDS stream to {:s}\n'.format(location))
+            # Create the rel_path under location
+            new_location = os.path.join(location, rel_path)
+            if not os.path.exists(new_location):
+                try:
+                    os.makedirs(new_location)
+                except OSError as e:
+                    sys.stderr.write('{:s}: {:s}\n'.format(e.strerror, new_location))
+                    continue
+            
+            move_status = True        
+            for f in f_contents:
+                # Move the file only if it doesn't already exist
+                dest_nc = os.path.join(new_location, os.path.basename(f))
+                if os.path.exists(dest_nc):
+                    sys.stderr.write('Destination file already exists: {:s}\n'.format(dest_nc))
+                    continue
+                try:
+                    sys.stdout.write('Moving {:s}: {:s}\n'.format(os.path.basename(f), new_location))
+                    shutil.move(f, new_location)
+                except IOError as e:
+                    sys.stderr.write('{:s}: {:s}\n'.format(e.strerror, f))
+                    move_status = False
+                    continue
+                    
+            if not move_status:
+                sys.stderr.write('Keeping stream source (1 or more move errors)\n')
+                continue
+                
+            # Prune the empty source directory if the move was successful
+            deleted_dirs = prune_empty_directories(TDS_NC_ROOT, rel_path)
+            for d in deleted_dirs:
+                sys.stdout.write('Deleted directory: {:s}\n'.format(d))
+            
+        elif args.delete:
             sys.stdout.write('Deleting THREDDS stream: {:s}\n'.format(tds_path))
             
             # Remove all files first
+            sys.stdout.write('Removing files...\n')
             for f in f_contents:
-                if not os.path.isfile(f):
-                    sys.stderr.write('Entry is not a file: {:s}\n'.format(f))
-                    continue
                 try:
                     sys.stdout.write('Deleting file: {:s}\n'.format(f))
                     os.remove(f)
@@ -124,34 +189,12 @@ def main(args):
                     continue
             # Recursively delete the directories from the bottom up provided they
             # are empty
+            sys.stdout.write('Pruning path: {:s}\n'.format(rel_path))
             deleted_dirs = prune_empty_directories(TDS_NC_ROOT, rel_path)
-            #d_tokens = rel_path.split('/')
-            #while d_tokens:
-            #    target_dir = os.path.join(*[TDS_NC_ROOT, os.path.join(*d_tokens)])
-            #    f_contents = os.listdir(target_dir)
-            #    if f_contents:
-            #        sys.stdout.write('Parent directory is not empty: {:s}\n'.format(target_dir))
-            #        break
-            #        
-            #    # Remove the empty directory
-            #    sys.stderr.write('Removing empty directory: {:s}\n'.format(target_dir))
-            #    os.rmdir(target_dir)
-            #    
-            #    # Remove the last element from d_tokens to go up one directory
-            #    d_tokens.pop(-1)
-
-                       
-        elif args.copy:
-            # Create the rel_path under args.location
-            new_location = os.path.join(args.location, rel_path)
-            try:
-                os.makedirs(new_location, mode=775)
-            except OS
-            sys.stdout.write('Copying THREDDS stream to {:s}\n'.format(args.location))
-        elif args.move:
-            sys.stdout.write('Moving THREDDS stream to {:s}\n'.format(args.location))
+            for d in deleted_dirs:
+                sys.stdout.write('Deleted directory: {:s}\n'.format(d))
     
-    status = 1
+    status = 0
               
     return status
 
@@ -164,13 +207,16 @@ def prune_empty_directories(root_dir, rel_path):
     d_tokens = rel_path.split('/')
     while d_tokens:
         target_dir = os.path.join(*[root_dir, os.path.join(*d_tokens)])
+        if not os.path.exists(target_dir):
+            sys.stderr.write('Directory does not exist: {:s}\n'.format(target_dir))
+            return deleted_dirs
+            
         f_contents = os.listdir(target_dir)
         if f_contents:
-            sys.stdout.write('Parent directory is not empty: {:s}\n'.format(target_dir))
-            break
+            sys.stdout.write('Directory is not empty: {:s}\n'.format(target_dir))
+            return deleted_dirs
             
         # Remove the empty directory
-        sys.stderr.write('Removing empty directory: {:s}\n'.format(target_dir))
         try:
             os.rmdir(target_dir)
         except OSError as e:
@@ -182,7 +228,11 @@ def prune_empty_directories(root_dir, rel_path):
         # Remove the last element from d_tokens to go up one directory
         d_tokens.pop(-1)
         
+    return deleted_dirs
+        
 def csv2json(csv_filename):
+    '''Parses the comma-separated value file, in which the first row is a column
+    header, and creates an array of dictionaries'''
     
     json_array = []
     
@@ -225,16 +275,11 @@ if __name__ == '__main__':
         help='Verbose display')
     arg_parser.add_argument('--tdsroot',
         type=str,
-        help='Location of the THREDDS root directory containing the source files')
+        help='Location of the THREDDS root directory containing the source files.  Must be specified if ASYNC_TDS_NC_ROOT is not set')
     arg_parser.add_argument('--location',
         type=str,
-        help='Location to move or copy directory tree')
-    arg_parser.add_argument('-v', '--validate',
-        action='store_true',
-        help='Validate that the directory tree exists, but perform no action')
-    arg_parser.add_argument('-x', '--debug',
-        action='store_true',
-        help='Debug mode.  Does not validate or perform operations')
+        help='Location to move or copy directory tree.  Must be specified if the --copy or --move option is used')
+
     parsed_args = arg_parser.parse_args()
 
     sys.exit(main(parsed_args))
